Isolate short job penalty types
